code/requirements.txt
dash
pandas
Pillow
matplotlib
keras
tensorflow
pycocotools
opencv-python